refactor(cijoe/bench): move shared functionality to imported script
The 'latency' and 'bench' workflows, which generate graphs and
reports, share functionality in the plotting and reporting paths. This
functionality could not be moved to a shared script until Cijoe added
support for importing self-made scripts in version 0.9.35.
