+++ cijoe/scripts/bench_reporter.py
@@ -14,41 +14,40 @@ Retargetable: True
 ------------------
 """
 from pathlib import Path
-from shutil import copyfile
 
 import jinja2
+from reporter import (
+    copy_graphs,
+    create_stylesheet,
+    create_test_setup,
+    create_xnvme_cover,
+    create_xnvme_info,
+)
 
 
 def main(args, cijoe, step):
     """Primary entry-point"""
-
     templates_path = Path(
         step.get("with", {}).get("templates", Path.cwd() / "templates" / "perf_report")
     ).resolve()
-
-    # Files emitted by this script
-    report_path = cijoe.output_path / "artifacts" / "perf_report"
-    cover_path = report_path / "cover.tmpl"
-    body_path = report_path / "report.rst"
-    style_path = report_path / "style.yaml"
-    pdf_path = report_path / "perf.pdf"
-
-    # Can be from a different run / results
-    prefix_path = Path(step.get("with", {}).get("path", cijoe.output_path)).resolve()
-    plot_path = prefix_path / "artifacts"
-
-    report_path.mkdir(parents=False, exist_ok=True)
-
-    # Fill 'report_path' with files needed to produce the .pdf
-    copyfile(templates_path / "style.yaml", style_path, follow_symlinks=False)
-
-    # Copy graphs from results/artifacts into 'report_path'
-    for png_path in plot_path.glob("*.png"):
-        print(png_path)
-        copyfile(png_path, report_path / png_path.name)
+    search_path = Path(step.get("with", {}).get("path", cijoe.output_path)).resolve()
+    artifacts = search_path / "artifacts"
 
     title = step.get("with", {}).get("report_title", "xNVMe")
-    subtitle = step.get("with", {}).get("report_subtitle", "Report")
+    subtitle = step.get("with", {}).get("report_subtitle", "Benchmark Report")
+
+    report_path = cijoe.output_path / "artifacts" / "perf_report"
+    report_path.mkdir(parents=False, exist_ok=True)
+
+    # Files emitted by this script
+    body_path = report_path / "report.rst"
+    pdf_path = report_path / "perf.pdf"
+    style_path = create_stylesheet(templates_path, report_path)
+    cover_path = create_xnvme_cover(templates_path, report_path, title, subtitle)
+    create_xnvme_info(templates_path, report_path)
+    create_test_setup(templates_path, report_path, artifacts)
+
+    copy_graphs(report_path, artifacts)
 
     template_loader = jinja2.FileSystemLoader(templates_path)
     template_env = jinja2.Environment(loader=template_loader)
